move more training data stuff into trainingdata directory
